Compile some CindyScript to JavaScript at build time
Now the drawgrid function resides in a separate file, and gets compiled to
JavaScript using the new cs2js tool.  That tool is very much designed to
cater for precisely those functions used by drawgrid right now, but it can
be easily extended in the future.  This is also a proof of concept for the
use of the Parser outside the main Cindy.js file, and for the compilation of
CindyScript to JavaScript which might be useful in other areas as well.

Before this change, drawgrid would not work since it violates the function
naming conventions.  Of course that could have been addressed by a rename.