Verify fuzzy comparison operators match their specification
This ensures that the fuzzy comparison operators EXACTLY match the
specification.  Since 10⁻¹⁰ can't be represented as a binary floating point
number, we use a mantissa-and-exponent computation in the reftest to exactly
describe the two closest approximations.  The actual value CSNumber.eps is
slightly larger than 10⁻¹⁰, so the comparison operations have to take that
into account.  Furthermore, to avoid loss of precision, all comparisons are
implemented based on the difference of the operands.