Extend set of identifier letters
This extends the set of characters permissible in identifiers
to the Unicode 8.0.0 category L, including astral planes.
The documentation already promises that this is what we use, but now
has a reftest to actually ensure astral planes work as advertised.

This change increases the size of the compiled code from 327048 to 328064
bytes before gzip, or 87248 to 87848 bytes after gzip.
Using the uncompressed regular expression literal would increase the size
after gzip by approximately 1500 bytes, even though it's less than 300 bytes
before gzip.