Improved language documentation
This now matches more closely what we have implemented, and what we intend.
Still not done yet, though.  In particular there are several details which
still need some clarification as to what we actually want there, before
documenting the choice we made.  It also might make sense to put the grammar
on a more formal level, in addition or instead of the informal description.