First stab at SVG rendering
This creates a “write to SVG” backend with an API that mimics
CanvasRenderingContext2D.  It can be used to export the current construction
to an SVG file, by typing the following command in the developer console:

    createCindy.instances[0].exportSVG()

The canvas support is very rough so far, and in particular there is no
support yet for any form of text rendering.  Raster images are unsupported
as well.  But it's a start.