Reworked API for dropped files
lastdropped now returns a List of Lists. Each list item consists of four
members: the value (image, string or nada), the CindyScript type, the MIME
type and the file name.  Stipping of file names was now left out at this
stage, since it's easy to strip that later on, while re-adding it later is
impossible.  The support for image as a ctype has been somewhat improved.