Add example for programmatic line position setting.
This adds an example which demonstrates the use of homog and dir field
setters for lines.  The code attempty to set all lines equal to the line PQ,
but for semi-free lines this obviously won't work in all cases.  The last
line is set via its angle, to ensure that this access works as well.