Update list of contributors in package.json
Switched to sorting by last name, and added Alexander Elkins.
The following command may be used to suggest perople to add:

    git shortlog -nes

It currently also mentions John E. Lloyd, the author of the quickhull
implementation. But since that's a separate project, included in a
subdirectory with its own COPYRIGHT file naming the author, I don't see him
as a contributor to CindyJS as such, but instead as the author of a
dependency of CindyJS which we chose to include in the repository.