Reduce number of assertion frameworks
Instead of using assert, should, expect and chai, we now only use the
different aspects of chai.  This will reduce the memory and download time
required for our dev dependencies, and should also make our test cases
slightly more consistent with one another.

The CSNumber test was written using expect but essentially was using
assertions, so I switched it to assertion style for brevity and clarity.