Introduce defaultAppearance.textsize
This text size is used as the default both for geometric objects (text
objects, point labels) and drawtext calls from CindyScript.  In CindyJS the
textsize defaults to 20.  This is different from the default of 12 used by
Cinderella, but since recent versions of Cinderella have been including
their default in the exported data, visual appearance should be preserved.