Flip sign of gravity
Since gravity is a quantity, not a vector, it may make more sense to
associate positive numbers with the more common “down” direction, even if
this is opposite to the geometric y axis.  That's what Cinderella does, so
this change of sign improves interoperability with Cinderella.  The code
handling the legacy syntax can take care of a legacy sign adjustment as
well, thus maintaining backwards compatibility.