Ensure we always have a div around our canvas
The new structure has a canvas inside a div which serves as a containing
block for the sake of absolute positioning within.  That way, HTML elements
can be positioned with respect to the canvas.

The HTML element identified by the port id (or port element or canvas or
canvasname) can be either the canvas element itself, or the div intended to
contain it.  The latter is in simpler since it requires less rewriting of
the DOM tree, so it is in fact the preferred way for new content.  The
example template already reflects that change.