Properly positioned push buttons
Since buttons are text in Cinderella, they get positioned like text.  For
this reason, it makes sense to treat them as text for now.  We do some case
distinctions pretty far down the line, in the actual text drawing handler,
where instead of drawing text to canvas we position the HTML element
according to the measured text dimensions.

It appears to be really hard to vertically align a HTML element based on
the baseline of its label text.  The best solution I could find was using
an image, aligning that to the baseline and ensuring that its known (and
preferrably large) height is the factor that determines the size of the
whole box.  Ugly, but it seems to work.
Found at http://stackoverflow.com/a/20445421/1468366 by keyboardSmasher.