Rewrite positioning code
Now we compute the width of the button label by measuring its HTML bounding
rectangle, instead of attempting to perform the same layout steps in the
canvas.  We also cache previous text properties, in order to reduce the
number of DOM modifications, particularly for buttons with fixed text, font
and position.