Reposition buttons if their size changes
This does catch the situation where the buttons get incorrectly positioned
at a time when the CSS stylesheet hasn't been loaded (or applied) yet.
However, currently this code only gets executed when drawing, which means
that for non-animated widgets we may still see an incorrect initial position
until some interaction triggers a repaint.  This can be reproduced using
Firefox with a full reload of the Buttons1.html example.