Improved horizontal positioning of buttons
Now the absolute position of a button can be set without knowing the width
of the button label.  The containing box has a negative margin to compensate
for border and padding, and alignment is achieved using a 2D transformation.

This should solve problems on Firefox where the relative offset between the
outer and inner bounding rectangles were computed incorrectly at first load
from an empty cache.