Clean up z-order
With the help of “display: inline-block”, we can turn the buttons into
blocks while still maintaining a baseline alignment.  We can then use
z-index to put the machinery for baseline alignment to the very back, the
canvas in front of that, and the buttons in front of the canvas.  That way,
the area above a button won't block access to the canvas or other buttons.