Rework update timing
Instead of calling updateCindy directly when something changed, we now make
sure that there is a requested animation frame which will run updateCindy.
This ensures that we have consistent call stacks when drawing, reducing the
risk of unexpected interactions between unrelated code.

This removes the D3 timer.  As far as I can tell from reading the code, all
it did in our setup was delegate to requestAnimationFrame, so we might as
well do that ourselves and avoid all that extra machinery we don't use.

This is a breaking change: the CindyScript `repaint` command no longer
causes an *immediate* execution of the `draw` script.  So code which
executes `repaint` and then relies on observing results generated in the
`draw` script would break.  This would be a pretty hackish approach, so most
sane widgets should be safe.