Maintain a generation counter for function definitions
This will allow code like CindyGL to detect when a given function has
changed.  The operator AST node is passed as a third argument to the
function implementing the execution of the operator, which can be used to
identify the code location but might also be useful for e.g. caching.