Show animation controls if requested
Show animation controls in an overlay in front of the canvas if the
`animcontrols` setting is set to true.

This commit cleans up CindyJS.scss somewhat to avoid excessive nesting and
to allow re-using CindyJS-specific classes outside the CindyJS widget area.

This is also the start for using SVG button icons.  The icons are served
from a single SVG file in order to allow sharing styles and other things
between icons, and to keep the number of transferred files low.  The
approach of using CSS and the :target selector to control layer visibility
was inspired by https://css-tricks.com/svg-fragment-identifiers-work/.