Handle SVG icon stack in JavaScript to keep Safari happy
Safari apparently does not support using the :target CSS property to select
the layer identified by the fragment part of a URL.  This breaks the use of
a SVG stack (i.e. a single SVG file containing multiple images) in the
common fashion.  On the other hand, having a SVG stack is quite convenient
for icon development since it allows easy sharing, and it also helps keep
number of HTTP requests low.  So this is an implementation of the SVG stack
idea in JavaScript.