Move animation settings to their own subobject
Now we have `animation.autoplay` and `animation.controls` (i.e.  properties
of a newly introduced `animation` subobject) instead of just `autoplay` and
`animcontrols` as options in the CindyJS invocation.
This adds structure, keeping related settings together.
It is expected that a new setting `speed` will be added to that object soon.