Support playtone via MIDI.js
The MIDI.js verison here is one which integrates three of my PRs:
https://github.com/mudcube/MIDI.js/pull/187 (percussion)
https://github.com/mudcube/MIDI.js/pull/189 (MIDI.now)
https://github.com/mudcube/MIDI.js/pull/190 (WebMIDI detection)

The playtone function automatically triggers loading of the plugin.  The
corresponding soundfont is loaded only when needed, which means there will
be some time between when the function calls start and when sound becomes
actually audible.

Samples are hosted on http://cindyjs.org/soundfont/ which might break
widgets delivered through HTTPS.  We need a certificate for cindyjs.org to
solve this problem, or allow configuring the soundfont through some API,
e.g. some CindyScript function.

The example is essentially taken from a Cinderella export, except for the
deletion of a duplicate play button.