Rewrite loop in loadScript for better readability
Beautify complained about this loop, and wanted to reformat it into a single
row.  While that was clearly worde than what we had, it indicated that the
loop might be unneccessarily hard to read.  So now instead of a for loop
with three lines, we have a while loop with one line before and one line in
the body, which should be easier to understand.

The whole point of this loop is to allow dot-separated identifiers as module
names.  This would be required if we were to use the sketch.ui.Timer to
report on sound font load progress.  Which we don't do at the moment.