Implement text operator to convert argument to string
In order to be compatible with Cinderella, the code does cast its argument
to a value if it is a geometric object.  This is in sync with current
Cinderella behavior, but the code there is annotated with a comment, so this
might change again in the future.

Contrary to Cinderella behavior, we always return a string, even if the
argument is nada.  In many cases this makes no difference, but in those
cases where it does make a difference, this behavior is arguably the more
useful one, so using it avoids extra case distinctions.