Allow multiple targets in create function
Now the create function allows for more than one element in the list of
names (first argument).  This is used in Cinderella for multi-valued
operations.  To model that in CindyJS, we create one element for the set of
results, and then a bunch of selection elements to pick from these.

As a drive-by fix, we now also handle geometric elements as arguments,
and allow algorithm names to name macros.