Reduce default line overhang to 1
The overhang of 1.1 will be useful only in a minority of situations, where
setting it explicitely would be the right thing to do.  In general people
clipping their lines will expect to see said lines drawn as segments with no
overhang.  This is also the default Cinderella uses, and which it assumes
when exporting, too.