Simplify segments and fix PointOnSegment
This simplifies segments in the following way: instead of storing three or
more points to establish a projective basis, we only store two, but we make
sure that we maintain a given relation between the representatives of these.
For a regular segment, the sum of these two endpoint vectors will be the
midpoint, while the difference will be the farpoint.  For projectively
transformed segments these relations hold for the images of the midpoint
resp. farpoint.

Besides the reduced number of properties, this is also essentially the
format which is most appropriate for PointOnSegment computations, since it
allows turning a parameter t into “start + t*(end - start)”, which the
previous implementation only did correctly for untransformed segments.

Another benefit is that it is now easier to figure out whether a given line
segment with real coordinates passes through infinity: If the z coordinates
of the endpoints have the same sign, the segment is finite, else it is
infinite.  This is far cheaper than a cross ratio computation would be.

This commit also edits examples to demonstrate and test new features.
In particular the 62_ProjTransform.html example was changed from a scripted
version (which cannot easily draw segments through infinity) to a GSLP-based
version which now gets rendered correctly.  Furthermore a point was added on
one of the projectively transformed lines, which maintains its position
relative to the grid even under projective transformations, demonstrating
the soundness of the new PointOnLine computation.
The example 15_ProjectiveGrid.html suffers from the same incorrect rendering
of line segments, but since it also demonstrates the use of map$8, it was
left as is in order to preserve the example for that function.
111_Moebius_Arc.html was modified to make the arcs more visible,
particularly in cases where they coincide with the grid.  This is needed to
ensure proper rendering of the segment-through-infinity case there.