Synchronize lab and simulation time
Now the time experienced by the physics lab and by the scripted animation
are in sync, accumulated in the per-instance variable simtime.  The time
gets updated in a mixed fashion: on devices capable of rendering 20fps or
more the sumulated amount of time is proportional to the actually elapsed
time, while for slower devices (or more complicated simulations) the
simulated amount of time is capped to the amount a 20fps rendering would
have between individual frames, so in this scenario the apparent simulation
speed is framerate-controlled and not walltime-controlled.

Cinderella uses a purely framerate-based approach.  Furthermore, it appears
as though Cinderella were using a different set of units internally, so that
for the same physical environment the same process would happen in vastly
different amounts of real and simulated time.  This commit compensates for
these differences using a pair of scale factors (simfactor and simunit)
which were determined empirically, using the simulation of a ball falling 12
units in a gravity of 0.01.  As the outcome for the wall clock timing will
depend on a number of factors, and in CindyJS even the amount of simulated
time will vary between runs due to different step sizes, these factors are
pretty rough estimates only.