Derive timing constants from Cinderella internals
After understanding Cinderella in more detail, an in particular the effect
of its “accuracy” setting, this code fixes the simunit constant to what it
is in the Cinderella codebase, and derives the simfactor constant from an
observation which is fairly easy to measure.
A long comment explains the details.

As another addition, the range of the speed control can now be set
arbitrarily.  This should be used in future Cinderella exports, and the
examples are already adjusted to that future representation.