Cinderella backward compatibility for speed
The speed setting of Cinderella (currently at build 1872) does not take the
accuracy (in the Cinderella but not CindyJS sense of the word) into account.
If the accuracy is at its default of 2, that means that the exported speeds
will be twice as much as what CindyJS should actually be using.  There
exists a merge request to export accuracy and also take it into account when
computing the speed.  Until then, we do a compatibility hack if speed is
exported, but accuracy is not.  The hack is limited to builds less than
1880, assuming that the export will be improved by then.