Support load$2 and load$3
The unary load$1 is not feasible in a web environment, since loading has to
happen asynchronously.  So we use a callback model: the function doing the
loading takes a piece of code to be invoked once loading has succeeded.  A
variable will contain the resulting value during execution of that code,
with the default of # as usual if the variable name is omitted.