Fix fullsize height for mobile
Apparently the height used for 100vh on mobile devices is the height of the
viewport if no browser controls are visible.  This is the case after the
page has been scrolled down.  However, since this kind of widget here is not
meant to be scrolled, we need the size which remains if the controls are
still visible.  Which we get using a height of 100%.

For details read https://github.com/bokand/URLBarSizing which has a good
summary of the current state of affairs in different browsers.