Demonstrate fullscreen widget
This demonstrates how a widget can be run in fullscreen mode if requested.
Particular care has to be taken with Safari, since there the CSS dimensions
will take precedence over the fullscreen request, possibly causing the
widget to get centered at its original size in a black screen.  To avoid
that we set the dimensions on an outer div, and size the inner div at 100%.
That way the fixed size will no longer apply in fullscreen mode.