Make text movable anywhere in its bounding box
For this to work we need to store the bounding box of the text when drawing
it.  This has to be supported by text rendering plugins like the KaTeX
plugin.  Since the object returned by the draw command is now a bounding box
instead of a CindyScript object, that function was moved to the helper
module and the drawtext$2 implementation now wraps around that.

As a drive-by fix, the selection offset for lines was changed from an offset
based in geometry coordinates to an offset based in canvas coordinates.  At
default zoom the two are the same, but the behavior should be less dependent
on choice of coordinate system now.