Capture key events from HTML overlay elements
This triggers the key event scripts also for HTML elements like buttons and
text boxes.  This is how Cinderella behaves.  Experiments indicate that this
does not lead to one key press triggering multiple executions of one of
these scripts.

One application of this change is code which responds to changed text using
such an event.  The “input” event would be more suitable for this situation,
but that doesn't have any counterpart in current Cinderella.  Key events do.