Make IFS buffer size a power of two
Buffer sizes for asm.js have to be at least 64 kiB, and either a power of
two or a multiple of 16 MiB.  We don't do a case distinction if buffer sizes
exceed 16 MiB, for the sake of simplicity.

See http://asmjs.org/spec/latest/#linking-0 for specification on this,
although Firefox doesn't seem to allow the 4 kiB lower limit that describes.