Draw few IFS in first iterations
This should lead to more responsive behavior, since the first results of an
operation will become visible more quickly.  Later on, when changes are
likely occurring more slowly, we stick to more points between updates as
there is considerable overhead associated with transferring the image to the
rendering pipeline in the main thread.