Avoid case distinctions and quadratic equation for PointOnArc
This drops one case distinction whether the defining points are collinear,
and another case distinction for the case of a line segment through
infinity.  This removes a lot of code reuse, so large portions of code were
moved back out of helpers to resemble what we originally had there.

The general style of PointOnArc, with initialize delegating to
getParamForInput, now closely resembles the corresponding structure for
PointOnSegment.

This also drops the quadratic equation (line-circle-intersection) in
PointOnArc.updatePosition.  Instead we now use the conicOtherIntersection
helper to obtain the point on the arc.  The idea is based on
  CR(A×I, C×I, B×I, P×I) = CR(A, C, B, P) = cr = crh_1 / crh_2 .
We can scale A×I and C×I such that their sum is a multiple of B×I to
establish a projective basis using these two representatives.  But we can
apply the same scaling factors to A and C, and obtain a point Q which lies
on both A×C and P×I.  Then we take the intersection of I×Q with the circle
that is different from I itself to find P.

The use of the conjugate here comes from the description of the
Moore-Penrose pseudoinverse in Wikipedia.  Although we don't compute the
pseudoinverse exactly that way, what we do is essentially equivalent, so
using the conjugate-transpose instead of the unconjugated transpose seems
the right thing to do here.  Not doing that leads to incorrect behavior
already in the initial state of the example file.