Point on Conic: project to nearest
This works reasonably well for moving the point with the mouse.  The
combination of orthogonal projection and Euclidean distance, discarding
complex points of intersection, will lead to a very intuitive placement of
the point at the nearest position on the conic.  This completely fails to
work if the conic moves from under the point, though, as the point will
become complex in this situation.  It's likely suboptimal for scripted
placement, too.