Only compute IFS parameters during rendering phase
As complex tracing will lead to complex intermediate steps, always
recomputing on updatePosition is a waste of resources.  It is better to only
re-initialize the IFS worker once we reached rendering and can expect real
coordinates.  It's an asynchronous operation anyway.

I had hoped that this would also avoid resetting the IFS2sim example in
response to the shift of the point E during the paint code, but that didn't
work out as hoped due to numeric errors.