Add apply, forall and sum to CindyGL
With this commit apply$2 and apply$3, forall$2 and forall$3 and sum$1 is implemented within colorplot.

This commit changes the handling of complex vectors: Now it is is literally a list of complex numbers instead of having both a real vector for the real and the complex part respectively. Probably the programs making use of complex matrix-vector multiplications get a little bit more slow as internal matrix-vector products cannot be used anymore.

A set of examples has also been included.

Under the hood, the expressions within apply or forall are unrolled. This is necessary as there is no way access entries of a list generically (as they are implemented as struct)