Spell out intended behavior for mover switching
This file demonstrates behavior in situations where a single mover gets
moved in different ways.  One possibility is moving the same object using
the mouse and using some script.  Another is moving the element in a series
of mouse movements, releasing the mouse button in between.

The file contains a list of suggested experiments along with intended
behavior.  We don't satisfy all these expectations yet, as releasing the
mouse button will still cause a checkpoint.