Report handling of tracingFailed situation
Now the tracingFailed flag is preserved even if the user releases the mouse
button.  One contribution here is the removal of stateContinueFromHere calls
from the mouse up handlers.  But even with that gone, the call to
traceMouseAndScripts would have caused a call to stateContinueFromHere if
there was no movement at all.  To avoid that, we have to preserve the
tracingFailed flag in such situations.  So we now clear that flag in two
cases: when we restore the stateLastGood (and thus throw away the failed
portion of the path) or when we call stateContinueFromHere (and thus make
the failed portion of the path permanent no matter what).  The latter can be
caused by a change in mover.