Draw conics in generic coordinate system
When drawing conics, do so in a coordinate system where the coordinate axes
don't line up with an asymptotic direction.  That avoids a large number of
special cases, including the one to distinguish between a non-degenerate
conic and a pair of lines, which in turn makes the results much more robust.