Connect boundary crossings for conic drawing
Now connecting the right boundary crossings works reasonably well.  At least
on the combinatoric side.  Actually drawing the lines is tricky for ellipses
where the shorter arc might be the one outside the boundary.  At this point
there is some code to draw ellipses as a whole, but that will likely break
for huge off-screen ellipses.  Needs some more work there.

The key ingredient in getting the combinatorics right is the following
consideration: consider a hyperbola which intersects the boundary in four
distinct points.  Then these are either two sections of the same branch, or
of different branches.  But how to tell?  A line connecting two points on
the same branch will intersect infinity outside the conic, while the line
connecting two points on different branches will intersect infinity inside
the conic.  If all boundary points are on the same branch, we want to span
the outside regions, while for points on different branches we want to span
the inside.  As the two signs cancel out, this concept makes sense even for
the degenerate case with a pair of lines.