Avoid recusing when ponts are already collinear
Without this, conics that degenerate into a pair of lines will require extra
recursions instead of drawing these lines straight from boundary crossing to
point of intersection in one call.  That's because while the point of
intersection for collinear tangents wasn't defined so far, all tangents of a
pair of lines pass through the point of intersection so this is now well
defined.  We still want to exit early.

This commit also gets rid of members of the point data structure which we no
longer need.