Handle conic through corners of viewport
If the conic passes (almost) through the corners of the viewport, then both
points of intersection for a given edge will be equally close to the center
of the edge.  Thus the distance from that center is not a suitable criterion
to decide which point to pick if exactly one is required.

On the other hand, if one solution is very far away, then using its sign may
lead to mistakes.  For this reason, the commit here uses both of these
concepts.  The decision is made based on whether or not the midpoint between
the two solutions is within the extent of the edge.  The switchover point
between the two code paths should be safely away from singular situations.