Allow supplying textures by object not name, allow NPOT textures
Now a texture can be given using any CindyJS-supported image format, not
only using a string by name.  There is also support for textures which are
not a power of two in size, rescaling them to make them power of two.
These two combine nicely to allow e.g. a live camera video as texture.