Compose transformations
This allows for the composition of transformations.  If a projective
transformation is composed with a Möbius transformation, there currently is
no check to ensure that the projective transformation is a Euclidean
transformation.  Ensuring this is the responsibility of the author at this
point, although a check may be added in the future.

Instead of a visual example there is a sizable test suite to ensure that
transforming a point by the transformations one after the other has the same
effect as applying the combined transformation.  The test framework was also
extended to verify some basic properties of Möbius transformations.