Ensure only Euclidean trafos get composed with Möbius ones
This prevents the composition of arbitrary projective transformatons with
Möbius transformations, as the result of these would not be a Möbius
transformation.  The property used to detect Euclidean transformations is
also used to keep track of their orientation-preserving or -reversing
nature, to avoid a non-analytic computation we had before.