Fix sign for Möbius IFS
It turns out that although the pair of IFS matrices generated for the
pixel-based transformation is correct, it no longer has the structure
created by moebiusPair so reading coefficients of these matrices is not a
valid operation.  Instead this approach here represents the view
transformation as a generalized Möbius transformation and uses the
established machinery to combine Möbius transformations.

There might be computationally cheaper solutions which take into account the
fact that some numbers are known to be real or even zero, and that the
resulting pair of matrices won't be needed at all.  But compared to the work
of actually rendering the IFS, setting up the pixel-based transformation is
cheap and reusing code eases maintainance.