Skip copying video frames of no new frames decoded
Before that, in every rendered frame video from a live-video stream was copied to the texture-Memory of the GPU. However, when dealing with huge resolutions (as a video stream from a spherical camera) often not so much data can be encoded.  That hopefully should prevent a lagging when dealing with huge video-streams.