Exclude suffix '.' from numbers if it could be a key.
A number in CindyJS can end with a dot, for instance, `2==2.`.
However, these the suffix dots should not always be parsed as be part of the number, because some succient dots have another meaning.
For instance, `1..5` is a list and is not to be parsed as `(1.).5`. (this has already been the case before this commit)
Also expressions `allpoints()_1.x` should be parsed as `(allpoints()_1).x` instead of `allpoints()_(1.)x`. (This is fixed with this PR).
In particular, JSON expressions such as `json = {"a":[1,{"b":2},3]}; json.a_2.b=34` should be valid.

With this commit, patterns `[number].[identifier]` lets the dot exclude from the number.